fix: get feat/work-order-nodes under check.lint.ui budget and refresh structure.sql
The retimestamped branch migrations bumped the `schema_migrations` version
recorded in `db/structure.sql` from main's `20260805201543` to our new highest
`20260805214133`. `check.db.structure` compares committed dump to the schema
a fresh migrate produces, so refresh the dump.

Also bring `check.lint.ui` back within budget (previously 600/593). Splits and
small refactors on files this branch added:

- Extract `<HeaderActions>` from `WorkOrderDetailHeader` — the function-body
  size was pushed over 120 lines by the new `Back to draft` guard.
- Extract `<WorkOrderCardDispatchButton>` + `<WorkOrderCardActors>` from
  `WorkOrderCard` to drop it under complexity 15.
- Extract `<ArtifactAttribution>` + `<ArtifactLabel>` from
  `ArtifactTimelineBody` to drop it under complexity 15.
- Extract `buildReadyState` from `useWorkOrderDetailPage` — the big return
  object was pushing it over complexity 15.
- Move the step-execution / dispatch-batch helpers to a new
  `workOrderTimelineStepBuilder.ts` so the main
  `workOrderTimelineFromEvents.ts` falls under 500 lines.
- Move the `*_WORK_ORDER_EVENTS` fixture constants and their helper
  functions to a new `factoryPageEventFixtures.ts` and update the two
  Storybook files that consume them.
- Pack the 6 positional args of `statusUpdatedEvent` into a `transition`
  options object (was tripping the max-params rule).
