+++ db/structure.sql
@@ -2681,7 +2681,7 @@ SET row_security = off;
 --
 
 COPY public.schema_migrations (version, dirty) FROM stdin;
-20260805201543	f
+20260805214133	f
 \.
