+++ web_src/src/pages/factories/WorkOrderDetailLoadedView.stories.tsx
@@ -5,24 +5,26 @@ import type { FactoriesWorkOrder, FactoriesWorkOrderArtifact, FactoriesWorkOrder
 import { ComponentStoryShell } from "./__fixtures__/ComponentStoryShell";
 import {
   CLOSED_FAILED_WORK_ORDER,
-  CLOSED_FAILED_WORK_ORDER_EVENTS,
   CLOSED_WORK_ORDER,
-  CLOSED_WORK_ORDER_EVENTS,
   DRAFT_WORK_ORDER,
-  DRAFT_WORK_ORDER_EVENTS,
   FACTORIES_ORGANIZATION_ID,
   FAILED_WORK_ORDER,
-  FAILED_WORK_ORDER_EVENTS,
   OPEN_WORK_ORDER,
   OPEN_WORK_ORDER_ARTIFACTS,
-  OPEN_WORK_ORDER_EVENTS,
   PRIMARY_FACTORY_ID,
   REFUND_FACTORY,
   REFUND_FACTORY_LINES,
-  RICH_OPEN_WORK_ORDER_EVENTS,
   RUNNING_WORK_ORDER,
-  RUNNING_WORK_ORDER_EVENTS,
 } from "./__fixtures__/factoryPageResponses";
+import {
+  CLOSED_FAILED_WORK_ORDER_EVENTS,
+  CLOSED_WORK_ORDER_EVENTS,
+  DRAFT_WORK_ORDER_EVENTS,
+  FAILED_WORK_ORDER_EVENTS,
+  OPEN_WORK_ORDER_EVENTS,
+  RICH_OPEN_WORK_ORDER_EVENTS,
+  RUNNING_WORK_ORDER_EVENTS,
+} from "./__fixtures__/factoryPageEventFixtures";
 import { WorkOrderDetailLoadedView } from "./WorkOrderDetailLoadedView";
 import { getWorkOrderDetailDerived } from "./lib/workOrderProgress";
